p1_lane_line_detection/requirements.txt
numpy==1.19.3
opencv-python
websockets
Pillow